requirements.txt
streamlit>=1.53.0
langchain>=1.2.0
langchain-community>=0.4.0
langchain-core>=1.2.0
openai>=1.0.0
databricks-sdk>=0.82.0
databricks-sql-connector>=4.2.0
python-dotenv>=1.0.0
